Update intro.md
- Check logs on pipeline overview should autoscroll you down to the logs - it's not intuitive to have to scroll down to see the logs. It kept making me think the button was broken.
The docs said the following - 
"This will take you to the orchestrate section, where you would need to select logs and check for logs corresponding to the start time of the relevant run. (ref. Step 8 in the orchestrate section above)."
The above is just blatantly wrong so I reworded it.

- Added new screenshots from dashboard (no notification system or llm analysis)